+++ src/interface.html
@@ -1,0 +1,91 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <title>AI Image Classifier</title>
+    <style>
+        body {
+            font-family: sans-serif;
+            display: flex;
+            flex-direction: column;
+            align-items: center;
+            justify-content: center;
+            min-height: 100vh;
+            background: #f8f9fa;
+        }
+
+        #preview {
+            width: 224px;
+            height: 224px;
+            object-fit: cover;
+            border: 2px solid #ccc;
+            margin-top: 10px;
+            border-radius: 12px;
+        }
+
+        button {
+            margin-top: 15px;
+            padding: 10px 20px;
+            border: none;
+            background: #007bff;
+            color: white;
+            font-size: 16px;
+            border-radius: 8px;
+            cursor: pointer;
+        }
+
+        button:hover {
+            background: #0056b3;
+        }
+
+        #result {
+            margin-top: 20px;
+            font-weight: bold;
+        }
+    </style>
+</head>
+
+<body>
+    <h2>Upload an Image for Classification</h2>
+    <input type="file" id="fileInput" accept="image/*">
+    <img id="preview" src="#" alt="Image preview" style="display:none;">
+    <button onclick="sendImage()">Predict</button>
+    <div id="result"></div>
+
+    <script>
+        const fileInput = document.getElementById('fileInput');
+        const preview = document.getElementById('preview');
+        const result = document.getElementById('result');
+
+        fileInput.addEventListener('change', () => {
+            const file = fileInput.files[0];
+            if (file) {
+                preview.src = URL.createObjectURL(file);
+                preview.style.display = 'block';
+            }
+        });
+
+        async function sendImage() {
+            const file = fileInput.files[0];
+            if (!file) {
+                alert("Please select an image first!");
+                return;
+            }
+
+            const formData = new FormData();
+            formData.append("file", file);
+
+            result.innerText = "Predicting...";
+
+            const response = await fetch("/predict/", {
+                method: "POST",
+                body: formData
+            });
+
+            const data = await response.json();
+            result.innerText = "Predicted class: " + data.predicted_class;
+        }
+    </script>
+</body>
+
+</html>
